// plugin/src/types/enum/chart-theme-types.ts
/**
 * Copyright 2024-present xxs3315
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export enum ChartThemeType {
    DEFAULT = 'default',
    VINTAGE = 'vintage',
    DARK = 'dark',
    WESTEROS = 'westeros',
    ESSOS = 'essos',
    WONDERLAND = 'wonderland',
    WALDEN = 'walden',
    CHALK = 'chalk',
    INFOGRAPHIC = 'infographic',
    MACARONS = 'macarons',
    ROMA = 'roma',
    SHINE = 'shine',
    PURPLE_PASSION = 'purple-passion',
    HALLOWEEN = 'halloween',
}